# Conflicts:
#	app/src/main/java/com/example/allgoing/fragment/CommunityFragment.kt
#	app/src/main/java/com/example/allgoing/fragment/ShopFragment.kt
#	app/src/main/res/layout/fragment_community.xml

+++ app/src/main/java/com/example/allgoing/retrofit/DTO/Response/StoreAllListRes.kt
@@ -10,7 +10,7 @@ data class StoreAllListRes(
         val storeIntro: String,
         val storeAddress: String,
         val star: Double,
-        val storeLatitude: Double,
-        val storeLongitude: Double
+        val storeLatitude: Int,
+        val storeLongitude: Int
     )
 }
